Stephen Agunsoye Resume
I am a graduate of Computer Science, with understanding of Computer programming and hands-on networking concepts. I am passionate about working with people, automating processes, and technology to provide guidance and implementation of security best practices. I am dedicated and ready for a career opportunity with a company that shares the same passion, vision, and values while providing growth opportunities for their team members.